Integrated the UI for adding a task with the UI for observing tasks. Created another controller to do so (and consequently renamed the original controller).
