+++ src/com/natebeckemeyer/projects/schedulrgui/graphics/AddTaskPopupController.java
@@ -1,29 +1,25 @@
 package com.natebeckemeyer.projects.schedulrgui.graphics;
 
-
-import com.natebeckemeyer.projects.schedulrgui.core.Parser;
 import com.natebeckemeyer.projects.schedulrgui.core.Schedulr;
 import com.natebeckemeyer.projects.schedulrgui.task.*;
 import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
-import javafx.scene.control.*;
-import javafx.scene.control.cell.PropertyValueFactory;
-import javafx.scene.input.KeyCode;
-import javafx.scene.input.KeyEvent;
+import javafx.scene.control.ChoiceBox;
+import javafx.scene.control.DatePicker;
+import javafx.scene.control.TextField;
+import javafx.stage.Stage;
 import javafx.util.StringConverter;
 
 import java.time.LocalDate;
 import java.util.GregorianCalendar;
 import java.util.LinkedList;
-import java.util.List;
 import java.util.Scanner;
 
-public class Controller
+/**
+ * Created for Schedulr by @author Nate Beckemeyer on 2016-06-09.
+ */
+public class AddTaskPopupController
 {
-    @FXML
-    private Button addTaskButton;
-
     @FXML
     private DatePicker addTaskDueDatePicker;
 
@@ -35,32 +31,6 @@ public class Controller
 
     @FXML
     private ChoiceBox<OnCompletion> addTaskChoiceBox;
-
-    @FXML
-    private TableView<Task> mainTaskList;
-
-    @FXML
-    private TextField taskListDefinition;
-
-    @FXML
-    private void provideView()
-    {
-        List<Task> passed = Schedulr.getTasksMatchingRule(Parser.processInput(taskListDefinition.getText()));
-        passed.sort(null);
-
-        ObservableList<Task> tasks = FXCollections.observableArrayList(passed);
-        mainTaskList.setItems(tasks);
-
-        // Due dates
-        @SuppressWarnings("unchecked")
-        TableColumn<Task, String> dueDates = (TableColumn<Task, String>) mainTaskList.getColumns().get(0);
-        dueDates.setCellValueFactory(new PropertyValueFactory<>("dueString"));
-
-        // Names
-        @SuppressWarnings("unchecked")
-        TableColumn<Task, String> names = (TableColumn<Task, String>) mainTaskList.getColumns().get(1);
-        names.setCellValueFactory(new PropertyValueFactory<>("name"));
-    }
 
     @FXML
     private void addTaskButtonClicked()
@@ -84,13 +54,7 @@ public class Controller
         newTask.setTags(tagList);
 
         Schedulr.addTask(newTask);
-    }
-
-    @FXML
-    private void enterTyping(KeyEvent key)
-    {
-        if (key.getCode() == KeyCode.ENTER)
-            provideView();
+        ((Stage) addTaskChoiceBox.getScene().getWindow()).close();
     }
 
     @FXML
@@ -115,4 +79,5 @@ public class Controller
 
         addTaskDueDatePicker.setValue(LocalDate.now());
     }
+
 }
