+++ src/com/natebeckemeyer/projects/schedulrgui/graphics/Main.java
@@ -8,36 +8,22 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
 
+import java.io.IOException;
+
 public class Main extends Application
 {
-
     @Override
-    public void start(Stage primaryStage) throws Exception
+    public void start(Stage primaryStage) throws IOException
     {
-        Parent root = FXMLLoader.load(getClass().getResource("addTaskPopup.fxml"));
+        Parent root = FXMLLoader.load(getClass().getResource("mainWindow.fxml"));
         primaryStage.setTitle("Schedulr");
         primaryStage.setScene(new Scene(root));
         primaryStage.show();
     }
 
-
     public static void main(String[] args)
     {
         Schedulr.addTasks(Parser.readTasksFromFile("tasks/current.dat"));
-
-        /*
-        Scanner console = new Scanner(System.in);
-
-        while (console.hasNextLine())
-        {
-            String input = console.nextLine();
-            if (input.equals("QUIT"))
-                return;
-
-            Schedulr.getTasksMatchingRule(Parser.processInput(input)).forEach(System.out::println);
-            System.out.println();
-        }*/
-
         launch(args);
     }
 }
