+++ src/com/natebeckemeyer/projects/schedulrgui/graphics/mainWindow.fxml
@@ -4,7 +4,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="579.0" prefWidth="599.0" xmlns="http://javafx.com/javafx/8.0.65"
             xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="com.natebeckemeyer.projects.schedulrgui.graphics.Controller">
+            fx:controller="com.natebeckemeyer.projects.schedulrgui.graphics.MainWindowController">
     <children>
         <TableView fx:id="mainTaskList" layoutX="-1.0" layoutY="53.0" prefHeight="526.0" prefWidth="600.0">
             <columns>
@@ -18,6 +18,7 @@
         <TextField fx:id="taskListDefinition" layoutX="97.0" layoutY="14.0" onKeyPressed="#enterTyping"
                    prefHeight="27.0" prefWidth="441.0" promptText="Define rule here!"/>
         <Button layoutX="545.0" layoutY="14.0" mnemonicParsing="false" onMousePressed="#provideView" text="View"/>
-        <Button layoutX="14.0" layoutY="14.0" mnemonicParsing="false" text="Add Task"/>
+        <Button layoutX="14.0" layoutY="14.0" mnemonicParsing="false" onMouseClicked="#viewAddTaskButtonClicked"
+                text="Add Task"/>
     </children>
 </AnchorPane>
